+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+= 1.0.2 =
+
+Fixed:
+
+* Duplicate weekly suggested tasks.
+* Fixed the REST API endpoint for getting stats.
+
 = 1.0.1 =
 
 Fixed: